% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles2rho.R
\name{angles2rho}
\alias{angles2rho}
\title{Matrix conversion tool}
\usage{
angles2rho(theta)
}
\arguments{
\item{theta}{A \eqn{(p-1)} times \eqn{(p-1)} matrix of angles.}
}
\value{
A \eqn{p} times \eqn{p} correlation matrix.
}
\description{
A tool for converting a matrix of angles to a correlation matrix using Choleksy factor
}
\examples{
theta <- matrix(rep(0.5, 4), 2, 2)
angles2rho(theta)

}
\seealso{
\code{\link[rcbmm]{rho2angles}}
}
