% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.step.1.R
\name{cm_step_1}
\alias{cm_step_1}
\title{Conditional-maximization step 1 (M-step 2)}
\usage{
cm_step_1(
  x,
  K,
  z,
  mvdc,
  margins,
  trace = TRUE,
  restrictions = TRUE,
  variance_tolerance = 1e-05
)
}
\arguments{
\item{x}{A numeric matrix or data frame of observations. Rows correspond to observations and columns correspond to variables.}

\item{K}{The number of mixture components.}

\item{z}{A numeric matrix representing the current value of the posterior probabilities of membership of the observations after the expectation step of the ECM algorithm. Columns are associated with a mixture component and rows are associated with observations. The current value of the probabilities should be computed via \code{\link[rcbmm]{e.step}}.}

\item{mvdc}{A list of objects of class \code{mvdc}. Each element of the list corresponds to a mixture component and contains the previous estimates for the component distribution's marginal and copula parameters.}

\item{margins}{A character vector specifying the marginal distributions of the components in the mixture. The vector must have a length equal to the number of columns in \code{x}. Each element must be equal to "norm", "beta" or "gamma".}

\item{trace}{A logical value indicating if an update regarding the step's progress should be displayed.}

\item{restrictions}{A logical value indicating if the variance of each Beta marginal should be restricted}

\item{variance_tolerance}{The lower bound for the variance of each Beta marginal, if \code{restrictions = TRUE}.}
}
\value{
A list of objects of class \code{mvdc}. Each element of the list corresponds to a mixture component and contains the updated estimates for the component distribution's marginal paramters and the same estimates as were parsed for the estimates of the copula parameter.
}
\description{
Implements the conditional-maximization step 1 of the ECM algorithm for regularized copula-based mixture model.
}
\seealso{
\code{\link[rcbmm]{cm.step.2}}
\code{\link[rcbmm]{ecm}}
}
