% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.step.R
\name{e_step}
\alias{e_step}
\title{Expectation step}
\usage{
e_step(x, K, mixing_probs, mvdc, margins)
}
\arguments{
\item{x}{A numeric matrix or data frame of observations. Rows correspond to observations and columns correspond to variables.}

\item{K}{The number of mixture components.}

\item{mixing_probs}{A numeric vector indicating the mixing proportions of the mixture model.}

\item{mvdc}{A list of objects of class \code{mvdc}. Each element of the list corresponds to a mixture component and contains the current estimates for the component distribution's marginal and copula parameters.}

\item{margins}{A character vector specifying the marginal distributions of the components in the mixture. The vector must have a length equal to the number of columns in \code{x}. Each element must be equal to "norm", "beta" or "gamma".}
}
\value{
A numeric matrix representing the posterior probabilities of membership of the observations after performing a single expectation step of the ECM algorithm. Columns are associated with a mixture component and rows are associated with observations.
}
\description{
Implements the expectation step of the ECM algorithm for regularized copula-based mixture model.
}
\seealso{
\code{\link[rcbmm]{ecm}}
}
