% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{ecm}
\alias{ecm}
\title{Model estimation through ECM}
\usage{
ecm(
  x,
  K,
  lambda,
  start = NULL,
  margins,
  trace = TRUE,
  maxit = 1000,
  epsilon = 1e-06,
  dist_mat = NULL
)
}
\arguments{
\item{x}{A numeric matrix or data frame of observations. Rows correspond to observations and columns correspond to variables.}

\item{K}{An integer specifying the number of components for which a regularized copula-based mixture model should be fitted.}

\item{lambda}{A numeric value indicating the value of the tuning parameter for regularization.}

\item{start}{A list providing the starting values for ECM. The list is produced by \code{\link[rcbmm]{fit.rcbmm}}.}

\item{margins}{A character vector specifying the marginal distributions of the components in the mixture. The vector must have a length equal to the number of columns in \code{x}. Each element must be equal to \code{"norm", "beta"} or \code{"gamma".}}

\item{trace}{A logical value indicating if an update regarding the algorithm's progress should be displayed after each iteration.}

\item{maxit}{A numeric value indicating the maximal number of ECM iterations.}

\item{epsilon}{A numeric value specifying the tolerance associated with determining when convergence of the ECM algorithm has been achieved.}

\item{dist_mat}{An object of type \code{dist} for calculating silhouette values.}
}
\value{
\itemize{
  \item{K}{The number of mixture components.}
  \item{lambda}{The value of the tuning parameter.}
  \item{z}{A numeric matrix representing the posterior probabilities of membership of the observations after the expectation step of the last iteration of the ECM algorithm. Columns are associated with a mixture component and rows are associated with observations.}
  \item{clusters}{A classification vector indicating the associated cluster of each observation. The classification corresponds to \code{z}.}
  \item{loglik}{A numeric vector displaying the penalized log-likelihood after each iteration of ECM.}
  \item{param_number}{The number of independent parameters associated with the model.}
  \item{BIC}{The BIC value of the model. Computed using the unpenalized log-likelihood after the last iteration of the ECM algorithm.}
  \item{mixing_probs}{The mixing proportions associated with the model.}
  \item{mvdc}{A list of objects of class \code{mvdc}. Each element of the list corresponds to a mixture component.}
  \item{transformation}{A character value indicating the transformation when identifying starting values. The value is \code{NULL} unless \code{lambda=0}. See \code{\link[rcbmm]{initialize.ecm}}.}
  \item{marginal_param}{A list containing the marginal parameters of the model as estimated by ECM. Each element corresponds to a mixture component.}
  \item{copula_param}{A list containing the copula parameters of the model as estimated by ECM. Each element corresponds to a mixture component.}
  \item{copula_param_angles}{A list containing the copula parameters of the model re-expressed as angles.}
  \item{silhouette}{See information regarding silhouette package. Add reference here.}
}
}
\description{
Implements the expectation-conditional-maximization algorithm for a regularized copula-based mixture model given initial parameter values, starting with the expectation step.
}
