% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.copula.pars.R
\name{extract.copula.pars}
\alias{extract.copula.pars}
\title{Extraction of copula parameters from a model}
\usage{
extract.copula.pars(mvdc, as_angles = F)
}
\arguments{
\item{mvdc}{A list of objects of class \code{mvdc}. Each element of the list corresponds to a mixture component and contains the estimates for the component distribution's marginal and copula parameters.}

\item{as_angles}{A logical value indicating whether the copula parameter should be returned as a matrix of angles instead of a correlation matrix.}
}
\value{
A list of matrices corresponding to the copula parameter of each mixture component contained in \code{mvdc}.
}
\description{
A function for extracting the copula parameters parameterizing the mixture components of a regularized copula-based mixture model from a list of \code{mvdc} objects. The copula parameter for each mixture component can be extracted in terms of a correlation matrix or a matrix of angles.
}
\seealso{
\code{\link[rcbmm]{extract.marginal.pars}}
}
