% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.rcbmm.R
\name{fit.rcbmm}
\alias{fit.rcbmm}
\title{Model selection}
\usage{
fit.rcbmm(
  x,
  lambda_grid,
  K = seq.int(2, 9),
  margins,
  maxit = 1000,
  epsilon = 1e-06,
  transform = TRUE,
  trace = FALSE
)
}
\arguments{
\item{x}{A numeric matrix or data frame of observations. Rows correspond to observations and columns correspond to variables.}

\item{lambda_grid}{An integer vector specifying the the values of the shrinkage parameter for which a regularized copula-based mixture model should be fitted. The default is \code{lambda_grid = 0}. If the vector parsed does not contain \code{0}, then \code{0} is appended as starting values can only be obtained in the case no regularization is applied to the model.}

\item{K}{An integer vector specifying the number of components for which a regularized copula-based mixture model should be fitted. The default is \code{K=2:9}.}

\item{margins}{A character vector specifying the marginal distributions of the components in the mixture. The vector must have a length equal to the number of columns in \code{x}. Each element must be equal to \code{"norm", "beta"} or \code{"gamma"}.}

\item{maxit}{A numeric value specifying the maximum number of iterations the ECM algorithm should run before being halted.}

\item{epsilon}{A numeric value indicting the tolerance for convergence.}

\item{transform}{A logical value indicating whether or not starting values for the case \code{lambda = 0} should be obtained using the transformations SPH, PCS, PCR and SVD. The default is TRUE.}

\item{trace}{A logical value indicating if an update regarding the step's progress should be displayed.}
}
\value{
\itemize{
  \item{BIC}{A matrix demonstrating the BIC values achieved by each model in the family. Each column corresponds to a given value of lambda and each row corresponds to a given number of components.}
  \item{SIL}{A matrix demonstrating the average silhouette width achieved by each model in the family. Each column corresponds to a given value of lambda and each row corresponds to a given number of components.}
  \item{all_models}{A list of lists with each element containing information about a specific model fitted.}
  \item{selected_models}{A list containing the information about the optimal model for each number of mixture components in the family. The optimal model for each number of componets is selected by picking the lambda that results in the largest average silhouette width. See the help file for silhouette for details.}
  \item{final_model}{The model contained in \code{selected_models} that maximized BIC.}
}
}
\description{
The function selects the most appropriate model from a family of regularized copula-based mixture models arising from a varying number of components and a differing shirnkage parameter.
}
\seealso{
\code{\link[rcbmm]{ecm}}
\code{\link[rcbmm]{initialize.ecm}}
}
