% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.ECM.R
\name{initialize.ecm}
\alias{initialize.ecm}
\title{Starting values for ECM}
\usage{
initialize.ecm(
  x,
  K,
  margins,
  transform = TRUE,
  hc_pairs = NULL,
  classification = NULL
)
}
\arguments{
\item{x}{A numeric matrix or data frame of observations. Rows correspond to observations and columns correspond to variables.}

\item{K}{The number of mixture componets.}

\item{margins}{A character vector specifying the marginal distributions of the components in the mixture. The vector must have a length equal to the number of columns in x. Each element must be equal to "norm", "beta" or "gamma".}

\item{transform}{A logical value indicating whether or not starting values should be obtained using the transformations SPH, PCS, PCR and SVD. The default is \code{TRUE}.}

\item{hc_pairs}{The results from MBHAC obtained from the function call \code{hc()} using the mclust package. If \code{NULL}, the function obtains the results by calling \code{hc()}.}

\item{classification}{A numeric vector representing a partitioning of the data \code{x}. If not \code{NULL}, the clustering identified by the vector takeß preference over the clustering identified by \code{hc_pairs}.}
}
\value{
\itemize{
  \item{mixing_probs}{A numeric vector indicating the starting values of the mixing proportions.}
  \item{mvdc}{A list of objects of class \code{mvdc}. Each element of the list corresponds to a mixture component and contains the starting values for the component distribution's marginal and copula parameters.}
  \item{transformation}{A character string indicating the transformation applied prior to performing MBHAC.}
  \item{loglik}{The log-likelihood corresponding of the model given the starting values of the parameters and the data.}
}
}
\description{
Identifies optimal starting values for ECM using the results of MBHAC applied to the data. The data undergoes transformations to enhance separation amongst groups prior to performing MBHAC.
}
\seealso{
\code{\link[mclust]{hc}}
}
