% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.martix.tools.R
\name{P2p.angles}
\alias{P2p.angles}
\alias{p2P.angles}
\title{Tool for handling angles}
\usage{
P2p.angles(theta)

p2P.angles(angles)
}
\arguments{
\item{theta}{A matrix of angles}

\item{angles}{A vector of angles}
}
\value{
A vector of angles

A matrix of angles
}
\description{
A tool for alternating between a matrix of angles and a vector of angles
}
