% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho2angles.R
\name{rho2angles}
\alias{rho2angles}
\title{Matrix conversion tool}
\usage{
rho2angles(rho_mat)
}
\arguments{
\item{rho_mat}{A \eqn{p} times \eqn{p} correlation matrix to be converted to a matrix of angles.}
}
\value{
A \eqn{(p-1)} times \eqn{(p-1)} matrix of angles.
}
\description{
A tool for converting a correlation matrix to a matrix of angles using Cholesky factor
}
\seealso{
\code{\link[rcbmm]{angles2rho}}
}
